Reorganize asset import style
