+++ admin-portal/public/EventCreator/EventCreator-index.tsx
@@ -1,0 +1,7 @@
+import logo from './images/logo.jpeg'
+import pen from './images/pen.png'
+
+export {
+    logo,
+    pen
+}
