+++ admin-portal/public/EventCreator/images/logo.jpeg
