+++ admin-portal/public/onboarding/images/create-filling.png
